# docs/api-gateway/swagger.yaml
basePath: /
definitions:
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingAggregateResponse:
    properties:
      booking:
        $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingResponse'
      payment:
        $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentResponse'
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingRequest:
    properties:
      check_in:
        type: string
      check_out:
        type: string
      guests:
        type: integer
      room_type_id:
        type: string
      user_id:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingResponse:
    properties:
      check_in:
        type: string
      check_out:
        type: string
      id:
        type: string
      status:
        type: string
      total_nights:
        type: integer
      total_price:
        type: number
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.CheckpointRequest:
    properties:
      action:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentResponse:
    properties:
      id:
        type: string
      payment_url:
        type: string
      provider:
        type: string
      status:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.ProfileResponse:
    properties:
      email:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
...
info:
  contact: {}
  description: Aggregated API surface for hotel booking platform.
  title: Hotel Booking Microservices API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens.
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /auth/me/{id}:
    get:
      description: Retrieve profile information for a user ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user with customer or admin role.
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      summary: Register user
      tags:
      - Auth
  /bookings:
    post:
      consumes:
      - application/json
      parameters:
      - description: Booking payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create booking
      tags:
      - Bookings
  /bookings/{id}:
    get:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get booking
      tags:
      - Bookings
  /bookings/{id}/cancel:
    post:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel booking
      tags:
      - Bookings
  /bookings/{id}/checkpoint:
    post:
      consumes:
      - application/json
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Checkpoint payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.CheckpointRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Booking checkpoint
      tags:
      - Bookings
  /bookings/{id}/status:
    post:
      consumes:
      - application/json
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      - description: Status payload
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      summary: Update booking status (internal)
      tags:
      - Bookings
  /gateway/aggregate/bookings/{id}:
    get:
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.BookingAggregateResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Aggregate booking data
      tags:
      - Gateway
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"